Fix transaction workspace exports
