+++ packages/transactional/package.json
@@ -5,7 +5,7 @@
   "type": "module",
   "license": "MIT",
   "exports": {
-    ".": "./src/index.ts"
+    ".": "./src/index.tsx"
   },
   "scripts": {
     "dev": "email dev --port 3002 --dir ./src/emails",
