+++ packages/transactional/tsconfig.json
@@ -3,6 +3,6 @@
   "compilerOptions": {
     "tsBuildInfoFile": "node_modules/.cache/tsbuildinfo.json"
   },
-  "include": ["*.ts", "src/emails", "src/index.ts"],
+  "include": ["*.ts", "src/emails", "src/index.tsx"],
   "exclude": ["node_modules"]
 }
